requirements.txt
flask
flask_sqlalchemy
mariadb